Dynamic car allocation and drawing
Each player is allocated now a random unique car on the server instead
of a random lane. The client render the current players cars.

Server game logic has major issue - when you lose life because you
bumped into an obstacle, you continue to lose life when on the next
turn, since move back with the obstacle. We need to either remove
obstacles or find another way to handle this.