Separate server and client models
Previously server and client were mixed, forcing installation of pygame
on the server for no good reason, and making the code harder to
understand (why server.matrix.py supports drawing?).

Now server has simple objects for manaing game state, and client has
simple objects for rendering game window. The server send the game
objects state to the client, which use the state to update the objects
that render the window.