Simple collision detection
Server ensure now that cars cannot take the same position on the track.
For fairness the server process each iteration in random order, giving
each player chance to take the first move. If a collision occurs, the
car is pushed back after the other cars or to the side.

We don't handle the case where car is pushed back over an obstacle.