Add example driver
This driver is scoring all available actions and choose the action with
the best score.

This driver prefer to keep in the same lane if possible since switching
lanes can lead to collisons with other cars. When it must switch lane,
he will take the lane that has a penguin a head.