Add XMLRPC interface for controling game server
Previously to start a game you had to start all the clients, then start
the server. The server would start the game when the first client
connected. To stop the server you had to stop the server process. This
caused players to get another car for each game and sometimes a player
could start one iteration after other players, depending on the timing
of starting the players.

Now you start and stop games using the new XMLRPC interface. A new
rtp-admin tool supports start and stop commands, and can be used
remotely.

Now players get a certain car when they join the game, and continue to
use the same car for all games until they disconnect.

As part of the xmlrpc interface, we have a web server. Currently it does
not serve any content, but it can be used to display the game
scoreboard.