Changebale game rate
Game rate was hardcoded in the server. Now it is configurable and
changeable while a game is running using the xmlrpc interface. Slowing
down the game is useful when you try to debug your driver, and making it
faster is useful for compraing drivers quickly.

To add the new verb, the rtp-admin tool was rewritten in a more
extensible way, so it will be eaiser to add more commands.