Simplify configuration and command line interface
Previously network configuration was in the rtp/common/config.py module,
which should not be edited by the students, and player name was given on
the command line.

Now the driver module has two additional required variables:

    server_address = "10.1.2.3"
    driver_name = "My Driver"

This configuration is typically done once and never changed, and the
command line interface is now simpler:

    ./rtp-client driver-module

One issue with the new configuration is that you have to duplicate a
file if you want to test several drivers using the same driver module.