Improve feedback when player join or leave the game
Previously client were update about the game state only after the game
was running. Now each player that join or leave the game send update to
all client.

This makes logging in the the game more clear. Instead of seeing an
empty track, you see your and others cars in the initial state.