Add Track.set() method
For testing, we need a way to set obstacles at non-random location and
check if the player action is handled correctly.

To unify track interface names, get_obstable() was renamed to
get() in Track and World.

Get method will be more correct when we expose other cars location using
World.get().