Add scoring tests
Scoring is the most tricky and badly implemented part of the game, and
we like to make significant logic changes. This patch adds tests
covering the current behavior, to make it easier to make the needed
changes later.

To make scoring more testable, scoring moved to a new module, so we do
not have to create a game object for testing scoring.