Fix scoring when turning into other player lane
We used to sort players by their in_lane status before processing
actions. This lead to possible wrong collision resolving, where a player
which is not in its lane *after* the turn won the collision.

Now we first handle RIGHT and LEFT actions, moving players into place,
and only then we sort players by in_lane status, and handle other
actions. If after the turn a player is not in its lane, it will lose a
collision with other player.