life calculations should be change when driving not in your lane #113
In the config.py I added a variable which supposed to punish the player when he takes a penguin, instead of gaining 10 points he gets 5.
In score.py I changed the code so when the player not in his lane, does't pass an obstacle he moves 2 cells backwards and punished by 10 points.