Draw the track background
The obstacles and the cars are not rendered yet.

This reveals an issue with the updater, we need to update in the same
rate of the game, but currently we update only once per second. The
track is rendered correctly only when using 1 FPS (the default).

Change the component interface to receive the entire state in the
update() method, since both the dashboard and the track need to know if
the game was started.