Use web socket for getting game updates
Replace the polling with web socket, receiving game state update when
the server broadcast updates to the players. Now the game is rendered
exactly like the python client.

Tweak the server update to send the last update after stopping, allowing
the client to change the user interface when the game ended normally.