Introduce the Hub class
Both Server and WatcherFactory are doing the same thing; adding and
removing clients, and broadcasting messages from the server to clients.

Extract a new Hub class that will add and remove clients and broadcast
messages from the server to clients, and add a simple PlayerFactory that
create a player with the hub.