Update clients when game is stopped
Previouly if a game was stopped from the command line admin tool, the
web admin status was not updated, since the server did not send an
update message with started: False. Change stop to send the last update
message so client are notified when game has finished, and do not need
to use the timeleft value.